% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColocalizeR.R
\name{ColocalizeR}
\alias{ColocalizeR}
\title{Colocalization analysis}
\usage{
ColocalizeR(x)
}
\description{
This function performs colocalization analysis on 2 channels of the user's choosing.
Relies on data generated by micimportR.
}
\examples{
colocalizeR()
}
\keyword{microscopR}

