% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_checker.R
\name{dataset_checker}
\alias{dataset_checker}
\title{Dataset checker}
\usage{
dataset_checker(x)
}
\description{
This function checks whether the dataset variable is present and if the user wants to remove it after processing.
Its sole purpose is a helper function that is called when necessary.
}
\keyword{microscopR}

