% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_remover.R
\name{dataset_remover}
\alias{dataset_remover}
\title{Dataset remover}
\usage{
dataset_remover(x)
}
\description{
This function follows dataset_checker and actually removes the dataset if requested.
Its sole purpose is a helper function that is called when necessary.
}
\keyword{microscopR}

