% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluoRatio.R
\name{fluoRatio}
\alias{fluoRatio}
\title{Ratiometric analysis}
\usage{
fluoRatio(x)
}
\description{
This function performs ratiometric analysis on 2 channels of the user's choosing.
Relies on data generated by micimportR.
}
\examples{
fluoRatio()
}
\keyword{microscopR}

