% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micimportR.R
\name{micimportR}
\alias{micimportR}
\title{Micrograph importer}
\usage{
micimportR(dir)
}
\arguments{
\item{dir}{root directory for image importing (does not support subdirectories)}
}
\description{
This function imports micrographs with EBImage and RBioFormats. It outputs a named list with 3 elements:
Data frames of the images, name of each image extracted from metadata and the base name of the project.
It is recommended to run one of the analysis functions after this.
}
\examples{
micimportR("/home/files/software/data/mouse-kidney")
}
\keyword{microscopR}

